requirements.txt
# Please note: you must install libshout3 and libshout3-dev before installing python-shout3 package!
# sudo apt-get install libshout3, libshout3-dev
# https://pypi.org/project/python-shout3/
python-shout3==0.2.6